Split segmentation.py into two modules
